pyproject.toml
[project]
name = "flowbridge"
version = "0.1.0"
description = "国内版 Zapier - 工作流自动化平台"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "tortoise-orm>=0.20.0",
    "aerich>=0.7.2",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "lark-oapi>=1.2.0",
    "jinja2>=3.1.0",
    "python-json-logger>=2.0.0,<3.0.0",
    "apscheduler>=3.10.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 127
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 127

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.aerich]
tortoise_orm = "src.conf.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
